objects/switch: merge Switch_Trigger in TRX
