+++ src/tr2/game/objects/general/switch.c
@@ -207,42 +207,6 @@ static void M_Control(const int16_t item_num)
     Item_Animate(item);
 }
 
-bool Switch_Trigger(const int16_t item_num, const int16_t timer)
-{
-    ITEM *const item = Item_Get(item_num);
-
-    if (item->object_id == O_SWITCH_TYPE_AIRLOCK) {
-        if (item->status == IS_DEACTIVATED) {
-            Item_RemoveActive(item_num);
-            item->status = IS_INACTIVE;
-            return false;
-        } else if (
-            (item->flags & IF_ONE_SHOT)
-            || item->current_anim_state == SWITCH_STATE_OFF) {
-            return false;
-        }
-
-        item->flags |= IF_ONE_SHOT;
-        return true;
-    }
-
-    if (item->status != IS_DEACTIVATED) {
-        return false;
-    }
-
-    if (item->current_anim_state == SWITCH_STATE_OFF && timer > 0) {
-        item->timer = timer;
-        if (timer != 1) {
-            item->timer = timer * LOGIC_FPS;
-        }
-        item->status = IS_ACTIVE;
-    } else {
-        Item_RemoveActive(item_num);
-        item->status = IS_INACTIVE;
-    }
-    return true;
-}
-
 static void M_SetupBase(OBJECT *const obj)
 {
     obj->control_func = M_Control;
